+++ home-automation-ui/src/App.js
@@ -7,6 +7,7 @@ import Guard from "./hoc/Guard/Guard";
 import RoomsDashboard from "./containers/RoomsDashboard/RoomsDashboard";
 import asyncComponent from "./hoc/asyncComponent/asyncComponent";
 import Login from "./containers/Auth/Login/Login";
+import Register from "./containers/Auth/Register/Register";
 
 function App() {
   const AsyncRoomsDevices = asyncComponent(() =>
@@ -17,6 +18,7 @@ function App() {
     <Layout>
       <Switch>
         <Route path="/login" exact component={Login} />
+        <Route path="/register" exact component={Register} />
         <Guard>
           <Route path="/" exact component={RoomsDashboard} />
           <Route path="/room/:id" exact component={AsyncRoomsDevices} />