+++ home-automation-ui/src/components/Auth/LoginBox/LoginBox.js
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import Button from "../UI/Button/Button";
-import Input from "../UI/Input/Input";
+import Button from "../../UI/Button/Button";
+import Input from "../../UI/Input/Input";
 
 import classes from "./LoginBox.module.scss";
 
@@ -50,12 +50,12 @@ export default class LoginBox extends Component {
             </div>
             <div className={classes.Row}>
               <div className={classes.Column}>
-                <Input placeholder={"Benutzername"} type={"text"} onChange={this.onChangeUsername}>Test</Input>
+                <Input placeholder={"Benutzername"} type={"text"} onChange={this.onChangeUsername}></Input>
               </div>
             </div>
             <div className={classes.Row}>
               <div className={classes.Column}>
-                <Input placeholder={"Passwort"} type={"password"} onChange={this.onChangePassword}>Test</Input>
+                <Input placeholder={"Passwort"} type={"password"} onChange={this.onChangePassword}></Input>
               </div>
             </div>
             <div className={classes.Row}>